♻️(front) capitalize component file names
Component names are capitalized. It only seems logical that the
files they are stored in have capitalized names too. It's also
standard practice.
