+++ richie/js/components/CourseGlimpse/CourseGlimpse.tsx
