+++ richie/js/components/CourseGlimpse/_CourseGlimpse.scss
