+++ richie/js/components/CourseGlimpseList/_CourseGlimpseList.scss
@@ -1,0 +1,3 @@
+.course-glimpse-list {
+  @include make-row();
+}
