+++ richie/js/components/Search/_Search.scss
