+++ richie/js/components/SearchFilter/_SearchFilter.scss
