+++ richie/js/components/SearchFilterGroup/_SearchFilterGroup.scss
