+++ richie/js/components/SearchFiltersPane/_SearchFiltersPane.scss
