--- richie/js/components/courseGlimpseList/_courseGlimpseList.scss
@@ -1,3 +1,0 @@
-.course-glimpse-list {
-  @include make-row();
-}
