+++ richie/static/scss/_main.scss
@@ -25,12 +25,12 @@
 @import "./partials/react-autosuggest";
 
 // Component styles
-@import "../../js/components/courseGlimpse/_courseGlimpse";
-@import "../../js/components/courseGlimpseList/_courseGlimpseList";
-@import "../../js/components/search/_search";
-@import "../../js/components/searchFilter/_searchFilter";
-@import "../../js/components/searchFilterGroup/_searchFilterGroup";
-@import "../../js/components/searchFiltersPane/_searchFiltersPane";
+@import "../../js/components/CourseGlimpse/_CourseGlimpse";
+@import "../../js/components/CourseGlimpseList/_CourseGlimpseList";
+@import "../../js/components/Search/_Search";
+@import "../../js/components/SearchFilter/_SearchFilter";
+@import "../../js/components/SearchFilterGroup/_SearchFilterGroup";
+@import "../../js/components/SearchFiltersPane/_SearchFiltersPane";
 
 // Template specific styles
 @import "../../../plugins/large_banner/templates/large_banner";
